extract the plp control hook
